+++ scripts/test-run-staging-integration-tests.sh
@@ -75,7 +75,8 @@ assert_runner_uses_manifest_transaction() {
     fi
     rg -q 'BYTEDEPTH_STAGING_IT_REDIS_PASSWORD' "$RUNNER"
     rg -q 'require_memory_headroom' "$RUNNER"
-    rg -q 'MINIMUM_MEMORY_AVAILABLE_KIB=524288' "$RUNNER"
+    rg -q 'MINIMUM_MEMORY_AVAILABLE_KIB=262144' "$RUNNER"
+    rg -q 'MINIMUM_MEMORY_AFTER_APP_STOP_KIB=524288' "$RUNNER"
 }
 
 assert_evidence_contract() {